requirements.txt
streamlit>=1.32.0
python-jobspy>=1.1.0
anthropic>=0.34.0
playwright>=1.44.0
apscheduler>=3.10.0
reportlab>=4.2.0
python-dotenv>=1.0.0
pandas>=2.2.0
requests>=2.31.0
Pillow>=10.0.0
groq>=0.9.0
